requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
deep-translator==1.11.4
Django==5.1.5
django-cors-headers==4.7.0
django-resized==1.0.3
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
